Rename {document,section}_association_marshaller*.rb
