+++ app/repositories/repository_registry.rb
@@ -1,7 +1,7 @@
 require "section_repository"
 require "manual_repository"
 require "section_edition"
-require "marshallers/document_association_marshaller"
+require "marshallers/section_association_marshaller"
 require "marshallers/manual_publish_task_association_marshaller"
 require "manual_publish_task"
 require "manual_with_publish_tasks"
